backend/requirements.txt
# ==================================================================================
# 🚀 CONVERSATIONAL RESEARCH DISCOVERY PLATFORM
# Hackathon: Elastic + Google Cloud AI Accelerate 2025
# ==================================================================================

# ==================================================================================
# CORE WEB FRAMEWORK
# ==================================================================================
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
starlette==0.48.0

# ==================================================================================
# ELASTICSEARCH (HYBRID SEARCH)
# ==================================================================================
elasticsearch==9.1.1
elastic-transport==9.1.0

# ==================================================================================
# GOOGLE CLOUD AI PLATFORM
# ==================================================================================
google-cloud-aiplatform==1.120.0
google-genai==1.42.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-cloud-bigquery==3.38.0
google-cloud-resource-manager==1.14.2
google-cloud-translate==3.21.1

# ==================================================================================
# GOOGLE DEPENDENCIES
# ==================================================================================
googleapis-common-protos==1.70.0
google-resumable-media==2.7.2
google-crc32c==1.7.1
grpcio==1.75.1
grpcio-status==1.75.1
grpc-google-iam-v1==0.14.2
proto-plus==1.26.1
protobuf==6.32.1

# ==================================================================================
# HTTP CLIENTS
# ==================================================================================
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.0
requests==2.32.5

# ==================================================================================
# DATA PROCESSING
# ==================================================================================
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1

# ==================================================================================
# TRANSLATION & LANGUAGE
# ==================================================================================
deep-translator==1.11.4
langdetect==1.0.9
googletrans==4.0.2

# ==================================================================================
# UTILITIES
# ==================================================================================
beautifulsoup4==4.14.2
numpy==2.3.3
pandas==2.3.3
networkx==3.5
tenacity==9.1.2

# ==================================================================================
# HTTP/2 SUPPORT
# ==================================================================================
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0

# ==================================================================================
# ASYNC SUPPORT
# ==================================================================================
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.11.0
uvloop==0.21.0
websockets==15.0.1

# ==================================================================================
# MISC DEPENDENCIES
# ==================================================================================
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
click==8.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
